Fix infrastructure.status.apiServerInternalURI for PublicAndPrivate
Currently in a PublicAndPrivate setup behind a proxy, nodes remain unready because ovn-node pods fail to talk to the control plane osdb and api endpoints.
This is because the connection to .hypershift.local tries to go through the proxy.
This PR set's the right value in the infra status that will result in hypershift.local being added to the NO_PROXY list of the ovn-pods via
https://github.com/openshift/cluster-network-operator/blob/8db9a00429dd4a5e4123b1217da93419557645eb/pkg/util/proxyconfig/no_proxy.go#L76
and
https://github.com/openshift/hypershift/blob/117c87f271ad7dc7cab875afb17769795fb7d117/control-plane-operator/controllers/hostedcontrolplane/cno/clusternetworkoperator.go#L248

Aditionally we'll also let the CNO set .hypershift.local in NO_PROXY unconditionally as this is a design invariant https://github.com/openshift/cluster-network-operator/pull/1692